Split logins sync functionality into separate modules
