+++ components/logins/src/sync/mod.rs
@@ -1,0 +1,32 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+mod engine;
+mod merge;
+mod payload;
+mod update_plan;
+
+use crate::error::*;
+pub use engine::LoginsSyncEngine;
+pub(crate) use payload::LoginPayload;
+
+#[derive(Clone, Copy, Debug, Eq, PartialEq, Ord, PartialOrd, Hash)]
+#[repr(u8)]
+pub(crate) enum SyncStatus {
+    Synced = 0,
+    Changed = 1,
+    New = 2,
+}
+
+impl SyncStatus {
+    #[inline]
+    pub fn from_u8(v: u8) -> Result<Self> {
+        match v {
+            0 => Ok(SyncStatus::Synced),
+            1 => Ok(SyncStatus::Changed),
+            2 => Ok(SyncStatus::New),
+            v => throw!(ErrorKind::BadSyncStatus(v)),
+        }
+    }
+}
